requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
pyyaml>=6.0.1
redis>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
